mixer library split into separate headers and name consistently
